BUG: non ccw polygon being interpreted
A workaround within cartopy is provided which changes the orientation
of a polygon if its exterior is of non counter-clockwise order.  The
order is important in cartopy while it isn't in matplotlib.  This is a
workaround for a bug present in mpl, as it is supposed to provide only
ccw polygons.