BUG: Fix incorrect ellipse boundary for Geostationary
Need to calculate boundary out to where the satellite viewing angle is
tangent to the spheroid (not out to +/- 90 latitude and longitude). Also
should always use a + h in the denominator, since this needs to be total
distance of the satellite above the earth center point.

Also go ahead and use numpy math functions since values are converted to
numpy floats.

This fixes a ring of GOES-16 data displaying outside the boundary
ellipse.