Use lengths to determine pcolormesh quads to wrap.
The current calculation of angle selects too many quads to use in the
pcolor call since it finds edges that go backward and *also* the edge
that returns forwards again.

Using a relative measure like the original means that the wrapping
pcolor will not be triggered if the initial pcolormesh is only a single
column.

The size is a bit arbitrary, but if pcolor is accidentally triggered for
such a large quad, the penalty is rather low compared to the penaltly
for several separate small quads.