Fix list parameters passed to artists
**Problem**

When artists are being drawn, the geometries outside the viewport
are removed before the drawing. The parameters that correspond to
removed geometries are left unchanged and so they are ill-matched
with the remaining geometries.

**Solution**

Keep track of which geometries are removed, and remove the
corresponding elements in the parameter values.

Fixes #580