Normalize grid coordinates used for interpolation.
Most coordinates are in metres and thus fairly large. This can cause
precision issues in Qhull as it calculates distances via squares and
square roots, at least on non-x86_64 architectures.

This is equivalent to the `rescale=True` argument in SciPy, but doing it
ourselves means a) we don't need to bump requirements to 0.14 and b)
re-scaling is not done for every scalar/vector property.