TST: Fix license checks with new git and Python.
New git-log does not accept empty strings for paths; we did not need to
pass an empty string since it was just an artifact of the parsing. Use
NUL-termination instead of newlines as this prevents git from adding any
additional porcelain to the filenames.

Also, use a raw-string for regexes to avoid invalid escapes.