Fix projected line strings getting truncated early.
The limit to prevent an infinite loop attempting to split line segments
"small enough" was applied to the entire line string, which meant the
entire path could get truncated. Instead, make the limit apply to each
segment individually. Since each segment is somewhat smaller, drop the
limit a bit to compensate for having to go through more segments. This
keeps things running in a reasonable time.