Turn GeoAxes.outline_patch into a GeoSpine.
This patch deprecates GeoAxes.outline_patch as well, since the spine
machinery works fine for what we want to do. The dictionary key is 'geo'
to match the one used by the Matplotlib GeoAxes.

The GeoSpine is a minimal implementation to have it work as it did
before, but more things may be implemented in the future if necessary.

Since PathPatch has default line width of 1.0, and spines have default
line width of 0.8, some boundary lines have changed slightly in the
test images.