Add an approx keyword arg to TransverseMercator.
Starting with Proj 6, the tmerc projection is implemented using the
former etmerc code. This is more accurate, but breaks our existing
expectations and tests.

The new keyword argument will, if True, instruct Proj6 to use the older
approximate code, and if False, the newer one. The default will change
in the next release.

Fixes #320.