Fix file paths for coverage on coveralls.
Currently, coverage is recorded for the file path as installed in the
conda environment on Travis. This causes coveralls to be unable to map
coverage to files in the repo, though it still produces final
statistics.

We need to flatten those paths back to the source directory, so that
coverage works properly.