Add a build envvar to force using Cython.
When running with tox, every install to the environment uses a single
sdist. This sdist will not contain the Cython-processed files (unless
you built them in your source tree previously), plus we want to rebuild
them anyway using the Cython in the environment.

This option, when added to tox configuration, allows the sdist to be
installed, even without pre-created Cython sources.