Set the default label formatter to lon/lat only for PlateCarree
Degrees have no meaning for other projections.

Set default formatter to ScalarFormatter when crs is not PlateCaree

Add also the xformatter and yformatter option to GeoAxes.gridlines:
the Gridliner is always initialized through GeoAxes.gridlines,
so it should be parameterized from there.

Note: The ScalarFormatter is the same as in MPL.