Improved the rotation of gridline labels
Labels are rectangles with text and their rotation make them
to overlap the map boundary, which force the current algorithm
to try another rotation angle, and eventually leads to suprising
rotations. This occurs because the labels are also moved along
an ellipse that is defined xpadding and ypadding.

This patch make the labels to move along the rectangle defined by xpadding
and yappding, the one that encloses the ellipse. This
sligthly increases the padding which may be sufficient to prevent
labels to overlap boundaries in most situations.