Update geoaxes.py
Update geoaxes pcolormesh: changed comment on diagonals
For axis projections other than Mercator and PlateCarree, compare the diagonals lenghts to the x extents in the y row, to take into account the shape of the plot in the detection of the big cells.
Update pcolor to catch a ValueError exception which happens for some big cells, by removing the artist instead of crashing the plot.