Use boundary to compute axes domain in gridliner
Fix #1443

Merge the old and the new algo for the axes domain estimate

The coordinates of the outline are simply added to the rasterised coordinates.

In addition, the rasterized coordinates are generated with a non-even number of ticks to make sure to catch the center when bounds are symetric.

Add gridliner_orthographic test

This tests Gridliner._axes_domain

Update figs due to new gridliner axes domain

Fix figure testing for crs

Fix tolerance for test_gridliner

Remove unneeded call to ravel