Changing the interpolation of vectors into the source projection
We now transform all source coordinate points into the target
projection and make a regular grid in the target projection. The
target grid points are then transformed back to the source
projection to do the interpolation in the source frame. Finally,
we transform the interpolated source grid vectors into the target
projection to return the fully transformed gridded data.