Updated tests to handle new get_array behavior with masking
A new test has been added which verifies that the data returned by
a pcolormesh get_array() command matches the data which was first
given to the pcolormesh command, even if wrapping has occurred.
One of the other tests (pcolormesh_diagonal_wrap) was also failing
due to a faulty check, which has been removed.