Correct Geostationary boundary on ellipses
When finding the tangent point, need to account for the ellipsoid in the
direction towards the satellite--which changes the trig. This is because
the line from the center is no longer perpendicular to the tangent line,
so we need some better math to figure out where that tangent point is.