BUG: Fix pcolormesh with nan inputs and wrapping
nan inputs to pcolormesh combined with a wrap would mask out the
data before getting to pcolor, and therefore the paths would not
be created. This creates the pcolor paths with generic data and then
calls set_array() with the proper data after the pcolor has been
created to make sure all paths are made first.