FIX: Change PlateCarree vector handling
If a PlateCarree projection is used for transforming vectors, users
likely have their data in longitude/latitude for the locations and
the measurement of the field is in North/South. To handle this
situation with PlateCarree requires an additional scaling by the
latitude of the point. This also requires some additional care
for handling points near the pole when applying this scaling.