MNT: Parameterize transform_vectors tests
This adds a parameterization for the vector transform tests on
the boundary.

The previous values were incorrect. They should be at an angle,
not all in vertical/horizontal coordinates. This can be thought of
by "wrapping" around the pole with a vector pointing up and right
at the corner (180, 90) turning into a down and left vector under
transform to stereographic.