Additional support for LINZ Maps Tiles
This is to add support for plotting NZ Map tiles, which is in EPSG:3857 (https://epsg.io/3857) already and in png format. Layer ID can be read in every layer's "About" tab, and the "subscription key" parameter is actually an API key given to the registered users. Tested in Python 3.8 and runs well.