FIX/TST: Use nearby boundary rather than extended segments
We don't "project" by a factor of a two to find the boundary
intersection, rather we project along the boundary to the nearest
point. So really we just want to make sure that our cuts are "close"
to the boundary, but we don't care about the segments being extended
by a given fraction.

This is relevant when the final two coordinates are close to each
other and thus that segment is not projected very far, yet it is
quite close to the boundary.