ci: Drop cibuildwheel matrix generation
We used to install cibuildwheel manually in order to generate the
Python-version matrix. This means that dependabot does not notice or
update it. This is normally fine, but when cibuildwheel adds new Python
versions, they won't be in the generated matrix until the pin is
updated.

Since our builds are relatively quick, drop the job matrix generation
and just split by platform/arch.

Also, expand the jq arguments so it's a bit clearer what it does, and
limit workflow permissions.