Add dummy renderer argument to _draw_preprocess
Commit 454567a2c8876f90969617be6fe4fd053f5e5e21 changes the method signature
of GeoAxes._preprocess_draw. This causes TypeErrors in user code that expects
draw to take a renderer argument, notably iris code with bbox_inches="tight".

This commit reinstates the argument in the _draw_preprocess function to make
the method signature compatible. The value of the argument is never used in
the function.