FIX: Gridliner should handle offset central longitudes
This merges parallel multi linestring geometries from an intersection
into a single linestring when possible. Before this, the line would start
in the middle of the axes and then go to the boundaries. The previous
logic would only take the start/end geometry points and would therefore
put the labels in the middle of the axes at "0" instead of on the edges.