Nightshade now works for extended longitudes
Now takes over=True switch, otherwise reverts to original
behaviour.

The addition of over=True to the RotatedPole Projection was
ultimately not useful here, but I've kept it for the moment in case
it is useful later (ideally all projections should accept this
switch, though I'm not sure it makes sense for Rotated Pole).