Extend longitude with Mercator projection; bug fix
The main modification is to enable the "over" parameter with
the Mercator projection.  This projection is not a subclass of
rectangular projections so required its own approach.  Most
things worked as intended, but stock_img required some extra
attention: in particular, when img_transform needed to be made
aware of the over parameter.  This wasn't necessary with Plate
Carrée because stock_img uses Plate Carrée internally already.

I also needed to modify how stock_img extended the image
because when transforming points they could be within numerical
error from zero, but the wrong side of it, messing up the
modulus function.